% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{catch_file_not_found}
\alias{catch_file_not_found}
\title{Catch file not found errors}
\usage{
catch_file_not_found(path, callback, ...)
}
\arguments{
\item{path}{path to read}

\item{callback}{IO function to use}

\item{...}{additional aruments passed to callback}
}
\value{
Returns the value of \code{callback(path, ...)} or NULL if a file not found error occured.
}
\description{
If an error occurs:
\itemize{
\item return NULL if it is due to a "path does not exist" error
\item stop otherwise
}
}
