% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fix_thousands_separator}
\alias{fix_thousands_separator}
\title{Remove thousands separator (',') in numbers
and convert to numeric.}
\usage{
fix_thousands_separator(col)
}
\description{
Remove thousands separator (',') in numbers
and convert to numeric.
}
