% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{parse_vcf_to_df}
\alias{parse_vcf_to_df}
\title{Parse VCF files for a provided path and construct data frame.}
\usage{
parse_vcf_to_df(path)
}
\arguments{
\item{path}{path to VCF file in \verb{*.vcf} or \verb{*.vcf.gz} format}
}
\value{
{tibble} new data frame with all variants (fixed field and genotype information)
}
\description{
Parse VCF files for a provided path and construct data frame.
}
