% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis}
\alias{read_personalis}
\title{Read a set of personalis results folders into a MultiAssayExperiment object}
\usage{
read_personalis(pathlist)
}
\arguments{
\item{pathlist}{List with paths to Personalis results folders}
}
\value{
MultiAssayExperiment
}
\description{
Read a set of personalis results folders into a MultiAssayExperiment object
}
