% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_bcr_clone_report_sample}
\alias{read_personalis_bcr_clone_report_sample}
\title{Read BCRs for a single sample.}
\usage{
read_personalis_bcr_clone_report_sample(sample_folder)
}
\value{
data frame with one row per BCR clone
}
\description{
Read BCRs for a single sample.
}
\keyword{internal}
