% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_bcr_isotype_composition_sample}
\alias{read_personalis_bcr_isotype_composition_sample}
\title{Read BCR isotype composition for a single sample}
\usage{
read_personalis_bcr_isotype_composition_sample(sample_folder)
}
\value{
data frame with isotype composition statistics
}
\description{
Read BCR isotype composition for a single sample
}
\keyword{internal}
