% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_cnv_report_sample}
\alias{read_personalis_cnv_report_sample}
\title{Read personalis CNV data for a single sample}
\usage{
read_personalis_cnv_report_sample(sample_folder)
}
\value{
data frame with all CNV
}
\description{
Read personalis CNV data for a single sample
}
