% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_cnv_reports}
\alias{read_personalis_cnv_reports}
\title{Read Personalis CNV data for a list of samples}
\usage{
read_personalis_cnv_reports(sample_paths)
}
\arguments{
\item{sample_paths}{List of directories with Personalis samples}
}
\value{
SummarizedExperiment
}
\description{
Read Personalis CNV data for a list of samples
}
