% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_gene_expression}
\alias{read_personalis_gene_expression}
\title{Read in gene expression data from personalis folders}
\usage{
read_personalis_gene_expression(sample_paths)
}
\arguments{
\item{sample_list}{A vector of paths to personalis folders}
}
\value{
SummarizedExperiment
}
\description{
Read in gene expression data from personalis folders
}
