% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_gene_expression_sample}
\alias{read_personalis_gene_expression_sample}
\title{Read a single gene expression sample from a personalis folder}
\usage{
read_personalis_gene_expression_sample(sample_folder)
}
\arguments{
\item{sample_folder}{path to personalis folder}
}
\description{
Read a single gene expression sample from a personalis folder
}
\keyword{internal}
