% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_hla_sample}
\alias{read_personalis_hla_sample}
\title{Read HLA types for a single sample}
\usage{
read_personalis_hla_sample(sample_folder)
}
\value{
data frame with one row per HLA gene per Allele.
}
\description{
Read HLA types for a single sample
}
\keyword{internal}
