% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_msi_reports}
\alias{read_personalis_msi_reports}
\title{Read MSI information for a list of samples}
\usage{
read_personalis_msi_reports(sample_paths)
}
\arguments{
\item{sample_paths}{Path to personalis samples}
}
\value{
SummarizedExperiment
}
\description{
MSI information is exclusively available form the HTML reports.
The information for individual loci will be stored in an assay of a SummarizedExperiment.
Summary statistics will be available from colData
}
