% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_msi_sample}
\alias{read_personalis_msi_sample}
\title{Extract MSI locus information and summary statistics from a single MSI HTML report}
\usage{
read_personalis_msi_sample(sample_folder)
}
\value{
SummarizedExperiment
}
\description{
Extract MSI locus information and summary statistics from a single MSI HTML report
}
