% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_small_variant_report_sample}
\alias{read_personalis_small_variant_report_sample}
\title{Read personalis Small Variant reports for a single sample.}
\usage{
read_personalis_small_variant_report_sample(
  sample_folder,
  modality,
  sample_type,
  report_type
)
}
\arguments{
\item{sample_folder}{path to personalis folder}

\item{modality}{modality from which the variants were called. Can be either 'DNA' or 'RNA'}

\item{report_type}{report_type which should be read. Can be either 'preferred' or 'all'}

\item{sample_types}{Can be one or multiple of of 'tumor', 'normal', or 'somatic'.
'tumor' refers to tumor sample vs. genome reference (i.e. somatic+germline mutations),
'normal' refers to normal sample vs. genome reference (i.e. germline mutations) and
'somatic' refers to tumor vs. normal (i.e. somatic mutations only).}
}
\value{
data frame with all variants
}
\description{
We only read in the "Preferred Transcript" report here:
}
\details{
Preferred Transcript: The RefSeq accession.version for the transcript used for variant analysis.
Personalis uses a curated list of transcripts, which is based on the number of
times a transcript (accession.version) is referred to in COSMIC. If not present in
COSMIC, the default transcript would be the one corresponding to the longest CDS.
(source: Personalis Analysis_Pipeline_Documentation)

We also do not read the \code{cancer_clinical_research}, the \code{cancer_research} and the \code{lowpupulationfreq} reports,
because they are subsets of the full report.

In addition, Personalis also provides raw VCF files with all (unfiltered) variants. If you need this level of information,
please specify read_vcf_data=TRUE in your read_personalis() call.
}
\keyword{internal}
