% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_small_variant_reports}
\alias{read_personalis_small_variant_reports}
\title{Read in small variant data from personalis folders}
\usage{
read_personalis_small_variant_reports(sample_paths, modality, sample_type)
}
\arguments{
\item{sample_paths}{A vector of paths to personalis folders}
}
\value{
SummarizedExperiment
}
\description{
Read in small variant data from personalis folders
}
