% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_tcr_clone_report_sample}
\alias{read_personalis_tcr_clone_report_sample}
\title{Read TCRs for a single sample.}
\usage{
read_personalis_tcr_clone_report_sample(sample_folder)
}
\value{
data frame with one row per TCR clone
}
\description{
Read TCRs for a single sample.
}
