% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_tcr_reports}
\alias{read_personalis_tcr_reports}
\title{Read Personalis TCR clone report for a list of samples}
\usage{
read_personalis_tcr_reports(sample_paths)
}
\arguments{
\item{sample_paths}{List of paths to read}
}
\value{
SummarizedExperiment
}
\description{
Read Personalis TCR clone report for a list of samples
}
