% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_tcr_summary_statistics}
\alias{read_personalis_tcr_summary_statistics}
\title{Extract summary statistics from the TCR HTML report}
\usage{
read_personalis_tcr_summary_statistics(sample_folder)
}
\value{
data frame
}
\description{
Extract summary statistics from the TCR HTML report
}
