% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_variant_calling_summary_statistics}
\alias{read_personalis_variant_calling_summary_statistics}
\title{Read HTML report on variant calling summary statistics and store information
based on given modality and sample_type:
DNA -> Normal DNA or Tumor DNA "Summary Small Variants" table
RNA -> Tumor RNA "Summary Small Variants" table
somatic -> "Tumor/Normal Concordance" table}
\usage{
read_personalis_variant_calling_summary_statistics(
  sample_folder,
  modality,
  sample_type
)
}
\arguments{
\item{sample_folder}{path to personalis folder}

\item{modality}{modality from which the variants were called. Can be either 'DNA' or 'RNA'}

\item{sample_type}{Can be one or multiple of of 'tumor', 'normal', or 'somatic'.
'tumor' refers to tumor sample vs. genome reference (i.e. somatic+germline mutations),
'normal' refers to normal sample vs. genome reference (i.e. germline mutations) and
'somatic' refers to tumor vs. normal (i.e. somatic mutations only).}
}
\description{
Read HTML report on variant calling summary statistics and store information
based on given modality and sample_type:
DNA -> Normal DNA or Tumor DNA "Summary Small Variants" table
RNA -> Tumor RNA "Summary Small Variants" table
somatic -> "Tumor/Normal Concordance" table
}
