% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_vcf_files_sample}
\alias{read_personalis_vcf_files_sample}
\title{Read Personalis (raw) VCF files for a provided folders/multiple samples and merge to one table.}
\usage{
read_personalis_vcf_files_sample(sample_folder, modality, sample_type)
}
\arguments{
\item{sample_folder}{base path to personalis folders}

\item{modality}{modality from which the variants were called. Can be either 'DNA' or 'RNA'}

\item{sample_type}{Can be one or multiple of of 'tumor', 'normal', or 'somatic'.
'tumor' refers to tumor sample vs. genome reference (i.e. somatic+germline mutations),
'normal' refers to normal sample vs. genome reference (i.e. germline mutations) and
'somatic' refers to tumor vs. normal (i.e. somatic mutations only).}
}
\value{
data frame with all variants
}
\description{
Read Personalis (raw) VCF files for a provided folders/multiple samples and merge to one table.
}
