% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{read_samples}
\alias{read_samples}
\title{Apply IO function for single sample to multiple samples}
\usage{
read_samples(sample_paths, io_func, description, ...)
}
\arguments{
\item{sample_paths}{{chracter} List of paths to samples}

\item{io_func}{{function} Callback function to apply to a single directory}

\item{description}{{character} Name of the modality that is read in (e.g. DNA). Used in the warning message
for missing samples.}
}
\value{
{list} List with one element per successfully read sample. Each element is a return value of \code{io_func}.
}
\description{
Ignores samples that cannot be found and prints a warning message for missing samples.
}
\keyword{internal}
