% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{warn_missing_samples}
\alias{warn_missing_samples}
\title{Print a warning message for each missing sample}
\usage{
warn_missing_samples(sample_list, sample_paths, mod_name)
}
\arguments{
\item{sample_list}{List of samples. Missing samples are represented by NULL}

\item{mod_name}{Name of the data modality to be used for the error message. e.g. "gene expression"}

\item{sample_path}{List of filenames from which the samples were read in (same order as sample_list)}
}
\value{
a boolean vector indicating which elements in \code{sample_list} are NULL
}
\description{
Print a warning message for each missing sample
}
