% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_small_variant_reports}
\alias{read_personalis_small_variant_reports}
\title{Read in small variant data from personalis folder
We only read in the "Preferred Transcript" report here:}
\usage{
read_personalis_small_variant_reports(sample_paths, modality, sample_type)
}
\arguments{
\item{sample_paths}{A vector of paths to personalis folders}

\item{modality}{modality from which the variants were called. Can be either 'DNA' or 'RNA'}

\item{sample_types}{Can be ome or multiple of of 'tumor', 'normal', or 'somatic'.
'tumor' refers to tumor sample vs. genome reference (i.e. somatic+germline mutations),
'normal' refers to normal sample vs. genome reference (i.e. germline mutations) and
'somatic' refers to tumor vs. normal (i.e. somatic mutations only).}
}
\value{
SummarizedExperiment
}
\description{
Preferred Transcript: The RefSeq accession.version for the transcript used for variant analysis.
Personalis uses a curated list of transcripts, which is based on the number of
times a transcript (accession.version) is referred to in COSMIC. If not present in
COSMIC, the default transcript would be the one corresponding to the longest CDS.
(source: Personalis Analysis_Pipeline_Documentation)
}
\details{
We also do not read the \code{cancer_clinical_research}, the \code{cancer_research} and the \code{lowpupulationfreq} reports,
because they are subsets of the full report.

In addition, Personalis also provides raw VCF files with all (unfiltered) variants. We currently don't
read them in because without additional filtering they are not very useful. If you need this level of information,
feel free to start from the raw vcf files or even run your own variant calling based on the FASTQ files.
}
