% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personalis.R
\name{read_personalis_vcf_files}
\alias{read_personalis_vcf_files}
\title{Read in (unfiltered) small variant data from VCF files from personalis folders}
\usage{
read_personalis_vcf_files(sample_paths, modality, sample_type)
}
\arguments{
\item{sample_list}{A vector of paths to personalis folders}
}
\value{
SummarizedExperiment
}
\description{
Read in (unfiltered) small variant data from VCF files from personalis folders
}
